+++ Sources/EmojiKit/_Deprecated/Emoji+Selection.swift
@@ -10,11 +10,8 @@ import Foundation
 
 public extension Emoji {
     
-    /// Register a user selection for the emoji.
-    ///
-    /// This will insert the selected emoji first within the
-    /// ``EmojiCategory/Persisted/recent`` category.
+    @available(*, deprecated, message: "Use the persisted emojis directly instead.")
     func registerUserSelection() {
-        EmojiCategory.Persisted.recent.addEmoji(self)
+        EmojiCategory.Persisted.frequent.addEmoji(self)
     }
 }